Changed sprites, added background, added code
Changed the size of the box and hero sprites, and added more variation.

Added background sprite taken from opengameart.org, and added copyright
information.

Added two wav files for warping in and out of the levels

Added code to the GameScene.swift file in order to add sprites and
context to the objects within.