Changed graphics stuff around
Changed around the way the sprites are added onto the scene by
implementing them in the .sks file instead of the .swift file.

Commented out all of the functioning code, and added in sections for
variables.

Also added in sections based on guidelines seen from numerous different
well-organized and structured code examples.

Provided some arrays to hold goal marker nodes and movable box sprites.

Added a variable to hold the information of the last touched position,
although I may want to opt for a d-pad style movement choice utilizing
tap-and-drag to move the character as opposed to a tap-to-move style. I
think the tap and drag style would suit the function of this game more
fully.