Added extra files and functional code
Added two new sound files for moving and contact

Removed the hero.swift file, as it was being unused

Added a d-pad visual to the game scene

Changed the sprites from being dependent upon code to display to
utilizing the GameScene.sks for sprite and game display.

Added functionality to the d-pad with a customized enumeration to
enable character movement.

Removed the old sprite display code

Added a function to update the display depending on where and what the
user is touching.

Added contact handling capabilities in accordance with Apple
documentation guidance in order to properly determine which physical
bodies are contacting one another.

Added a move action function, which would take the distance, sprite
information, and speed to add movement based on linear interpolation.