Fixed sound loading
Fixed the loading of sound, moving the sound files to be loaded prior
to the running of the scene itself, as per guidance of course
instructor.

Added an end level function to build the win/lose condition.

Working on utilizing the built in update function to further utilize
the functionality intended for linear interpolation.