Finished Beta Section
Modified the info.plist to handle iPad rotation to handle possible
graphics issues.

Added resizing of onscreen labels and controls depending upon which
type of device is being used.

Added a new atlas file with an animation reference created using
TexturePacker.

Modified the view controller to return .Landscape for all devices.

Reformatted some of the code to make it easier to read.

Declared audio outside of the view to provide better memory allocation.

Also declared nodes once on loading the view, instead of declaring them
inside functions in order to prevent memory leaks.

Added an animation function/action

Added a blocking sprite for a win/lose condition in order to provide
further user feedback as well as preventing the user from interacting
with anything onscreen aside from a restart button.

Set a win and lose condition.

Added a scaling factor for the right side bar, which changes depending
on the type of device being used.

Added a score tracker, as well as a move tracker.

Enhanced the physics of the game by removing some of the inherent
“squish” factor of the hero.

Utilizing the update function to check for the lose condition.

Added a pause button, and ensured proper functionality.

Provided a restart function that reloads the current scene in the same
manner that the scene is originally loaded.