Initial Commit for IAD
Initial commit for IAD course

Changes:

Changed the original base scene swift file into it’s own class type in
order to function as a general engine behind the function of the game
itself.

Created a new file for the first level of the game, as a class type of
the BaseScene.

Changed the function of the MenuScene file to reflect the change in
classes.