Added GameKit functionality, leaderboards
Added GameKit functionality

Provided Singleton for Game Center processing purposes, also declared a
global Score variable in the singleton file.

Added a new section on the menu scene to access the Leaderboards
created in iTunes Connect

Added coding to corresponding MenuScene.swift file to show the
leaderboard in the appropriate kind of View Controller.

Added the save high score function into the Singleton, so it will be
accessible across multiple areas of the application.

Important note: Xcode automatically force-updated to the next
generation. Due to this change, some of the functionality of the
original application is not performing as it was originally when
running the code in Swift 2.3. The code has been left as-is in the
interest of performance standards for Swift 2.2.