Incorporated achievements
Added achievement processing to the Singleton file.

Added the ability to save the number of powerups activated to the
device defaults, including the function in the termination portion of
the App Delegate

Added achievements to measure moves, power ups iteration, losing, and
winning.

Ensured that the achievements show a banner on receipt.